pyproject.toml
[tool.poetry]
name = "gcsfs-aoi"
version = "0.1.0"
description = ""
authors = ["Noah D. Brenowitz <nbren12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pylint = "^2.4.4"
aiohttp = "^3.6.2"
gcloud-aio-storage = "^5.2.2"
gcsfs = "^0.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
black = "^19.10b0"
zarr = "^2.4.0"
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
